+++ StoreApp/Podfile
@@ -1,0 +1,19 @@
+# Uncomment the next line to define a global platform for your project
+# platform :ios, '9.0'
+
+target 'StoreApp' do
+  # Comment the next line if you don't want to use dynamic frameworks
+  use_frameworks!
+	pod 'Toaster'
+  # Pods for StoreApp
+
+  target 'StoreAppTests' do
+    inherit! :search_paths
+    # Pods for testing
+  end
+
+  target 'StoreAppUITests' do
+    # Pods for testing
+  end
+
+end
