+++ StoreApp/StoreApp/ViewController.swift
@@ -7,6 +7,7 @@
 //
 
 import UIKit
+import Toaster
 
 class ViewController: UIViewController, UITableViewDelegate {
     @IBOutlet weak var tableView: UITableView!
@@ -17,6 +18,7 @@ class ViewController: UIViewController, UITableViewDelegate {
         super.viewDidLoad()        
         tableView.delegate = self
         tableView.dataSource = menuModel
+        Toast(text: "hi").show()
     }
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
         return 140
@@ -24,12 +26,15 @@ class ViewController: UIViewController, UITableViewDelegate {
     func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
         let header = tableView.dequeueReusableCell(withIdentifier: "HeaderCell") as! HeaderCell
         let headerData = menuModel.menuList[section]
-        header.category?.text = headerData.category ?? ""
-        header.title?.text = headerData.title ?? ""
+        header.setData(title: headerData.title, category: headerData.category)
         return header
     }
     func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
         return 80
+    }
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        let item = menuModel.menuList[indexPath.section][indexPath.row]
+        Toast(text: "\(item?.title) : \(item?.s_price)").show()
     }
 }
